prototypes/rag-demo-app/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.5.1
PyMuPDF==1.23.26
python-docx==1.1.0
pandas==2.2.1
openpyxl==3.1.2
Pillow==10.2.0
werkzeug==3.0.1
numpy==1.26.4
tqdm==4.66.2
